ai-damage-detector/app.py
import cv2
import numpy as np
from tensorflow.keras.models import load_model
import os
from azure.storage.blob import BlobServiceClient

# Configuration
class Config:
    """Configuration class for the damage detector"""
    MODEL_PATH = "model/damage_detection_model.h5"
    IMAGE_SIZE = (256, 256)
    DAMAGE_CLASSES = ['low', 'moderate', 'severe']
    
    # Azure Storage configuration
    CONNECTION_STRING = os.getenv('AZURE_STORAGE_CONNECTION_STRING')
    CONTAINER_NAME = 'disaster-images'

class DamageDetector:
    """Main class for damage detection"""
    def __init__(self):
        """Initialize the damage detector"""
        self.model = load_model(Config.MODEL_PATH)
        self.blob_service_client = None
        
        if Config.CONNECTION_STRING:
            self.blob_service_client = BlobServiceClient.from_connection_string(Config.CONNECTION_STRING)

    def process_image(self, image_path):
        """
        Process an image and predict damage severity
        
        Args:
            image_path (str): Path to the image file
            
        Returns:
            dict: Prediction results including severity and confidence
        """
        # TODO: Implement image processing and prediction
        pass

    def upload_to_azure(self, image_path, metadata=None):
        """
        Upload processed image to Azure Blob Storage
        
        Args:
            image_path (str): Path to the image file
            metadata (dict): Optional metadata for the blob
        """
        if not self.blob_service_client:
            return
            
        # TODO: Implement Azure Blob Storage upload
        pass

if __name__ == "__main__":
    detector = DamageDetector()
    # TODO: Implement main processing loop
