ai-damage-detector/requirements.txt
opencv-python==4.8.0
numpy==1.24.3
tensorflow==2.13.0
azure-storage-blob==12.20.0
pillow==10.0.0
scikit-learn==1.3.0
matplotlib==3.7.1
